modified the add nic operation.
