+++ cs-nutanix-prism/src/main/java/io/cloudslang/content/nutanix/prism/services/VMImpl.java
@@ -13,23 +13,15 @@
  * limitations under the License.
  */
 
-package io.cloudslang.content.nutanix.prism.service;
+package io.cloudslang.content.nutanix.prism.services;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.cloudslang.content.httpclient.entities.HttpClientInputs;
 import io.cloudslang.content.httpclient.services.HttpClientService;
-import io.cloudslang.content.nutanix.prism.entities.NutanixAddNicInputs;
-import io.cloudslang.content.nutanix.prism.entities.NutanixCreateVMInputs;
-import io.cloudslang.content.nutanix.prism.entities.NutanixGetVMDetailsInputs;
-import io.cloudslang.content.nutanix.prism.entities.NutanixListVMsInputs;
-import io.cloudslang.content.nutanix.prism.service.models.virtualmachines.AddNicRequestBody;
-import io.cloudslang.content.nutanix.prism.service.models.virtualmachines.CreateVMRequestBody;
-import io.cloudslang.content.nutanix.prism.utils.Constants;
 import io.cloudslang.content.nutanix.prism.entities.*;
-import io.cloudslang.content.nutanix.prism.service.models.virtualmachines.CreateVMRequestBody;
-import io.cloudslang.content.nutanix.prism.service.models.virtualmachines.SetVMPowerStateRequestBody;
-
+import io.cloudslang.content.nutanix.prism.services.models.virtualmachines.CreateVMRequestBody;
+import io.cloudslang.content.nutanix.prism.services.models.virtualmachines.SetVMPowerStateRequestBody;
 import org.apache.http.client.utils.URIBuilder;
 import org.jetbrains.annotations.NotNull;
 
@@ -38,7 +30,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
-import static io.cloudslang.content.nutanix.prism.service.HttpCommons.setCommonHttpInputs;
+import static io.cloudslang.content.nutanix.prism.services.HttpCommons.setCommonHttpInputs;
 import static io.cloudslang.content.nutanix.prism.utils.Constants.Common.*;
 import static io.cloudslang.content.nutanix.prism.utils.Constants.GetVMDetailsConstants.GET_VM_DETAILS_PATH;
 import static io.cloudslang.content.nutanix.prism.utils.Constants.SetVMPowerStateConstants.SET_POWER_STATE_PATH;
@@ -171,79 +163,6 @@ public class VMImpl {
                 .append(API)
                 .append(nutanixCreateVMInputs.getCommonInputs().getAPIVersion())
                 .append(GET_VM_DETAILS_PATH);
-        uriBuilder.setPath(pathString.toString());
-        return uriBuilder.build().toURL().toString();
-    }
-
-    @NotNull
-    public static Map<String, String> AddNic(@NotNull final NutanixAddNicInputs nutanixAddNicInputs)
-            throws Exception {
-        final HttpClientInputs httpClientInputs = new HttpClientInputs();
-        httpClientInputs.setUrl(AddNicURL(nutanixAddNicInputs));
-        setCommonHttpInputs(httpClientInputs, nutanixAddNicInputs.getCommonInputs());
-        try {
-            httpClientInputs.setBody(AddNicBody(nutanixAddNicInputs));
-        } catch (JsonProcessingException e) {
-            return getFailureResultsMap(e);
-        }
-        httpClientInputs.setAuthType(BASIC);
-        httpClientInputs.setMethod(POST);
-        httpClientInputs.setUsername(nutanixAddNicInputs.getCommonInputs().getUsername());
-        httpClientInputs.setPassword(nutanixAddNicInputs.getCommonInputs().getPassword());
-        httpClientInputs.setContentType(APPLICATION_API_JSON);
-        return new HttpClientService().execute(httpClientInputs);
-    }
-
-    @NotNull
-    public static String AddNicBody(NutanixAddNicInputs nutanixAttachNicInputs)
-            throws JsonProcessingException {
-        String requestBody = EMPTY;
-        ObjectMapper AttachNicMapper = new ObjectMapper();
-        AddNicRequestBody AddNicBody = new AddNicRequestBody();
-        ArrayList spec_list = new ArrayList();
-        String[] network_uuid = nutanixAttachNicInputs.getNetworkUUID().split(",");
-        String[] requested_ip_address = nutanixAttachNicInputs.getRequestedIPAddress().split(",");
-        String[] is_connected = nutanixAttachNicInputs.getIsConnected().split(",");
-        String[] vlan_Id = nutanixAttachNicInputs.getVlanid().split(",");
-        if (requested_ip_address.length == is_connected.length && requested_ip_address.length == vlan_Id.length) {
-            for (int i = 0; i < requested_ip_address.length; i++) {
-                AddNicRequestBody.VMNics vmNics = AddNicBody.new VMNics();
-                vmNics.setNetwork_uuid(network_uuid[i]);
-                vmNics.setRequested_ip_address(requested_ip_address[i]);
-                vmNics.setIs_connected(Boolean.parseBoolean(is_connected[i]));
-                vmNics.setVlan_id(vlan_Id[i]);
-                spec_list.add(vmNics);
-            }
-        } else {
-            for (int i = 0; i < requested_ip_address.length; i++) {
-                AddNicRequestBody.VMNics vmNics = AddNicBody.new VMNics();
-                vmNics.setNetwork_uuid(network_uuid[i]);
-                vmNics.setRequested_ip_address(requested_ip_address[i]);
-                spec_list.add(vmNics);
-            }
-        }
-        AddNicBody.setSpec_list(spec_list);
-
-        try {
-            requestBody = AttachNicMapper.writeValueAsString(AddNicBody);
-        } catch (JsonProcessingException e) {
-            e.printStackTrace();
-        }
-
-        return requestBody;
-    }
-
-    @NotNull
-    public static String AddNicURL(NutanixAddNicInputs nutanixAttachNicInputs) throws Exception {
-
-        final URIBuilder uriBuilder = getUriBuilder(nutanixAttachNicInputs.getCommonInputs());
-        StringBuilder pathString = new StringBuilder()
-                .append(API)
-                .append(nutanixAttachNicInputs.getCommonInputs().getAPIVersion())
-                .append(GET_VM_DETAILS_PATH)
-                .append(PATH_SEPARATOR)
-                .append(nutanixAttachNicInputs.getVmUUID())
-                .append(Constants.AddNICConstants.ADD_NIC_PATH);
         uriBuilder.setPath(pathString.toString());
         return uriBuilder.build().toURL().toString();
     }
